fixed real quick

+++ events/message.js
@@ -5,7 +5,9 @@ module.exports = msg => {
       return msg.reply('Can you stop spamming mentions so I can go back to my eucalyptus?');
     }
     if (msg.content.includes('@everyone') || msg.content.includes('@here')) {
-      msg.delete();
-      msg.reply('I\'m trying to eat eucalyptus in peace right now, so please stop trying to tag everyone.');
+      if (!msg.member.hasPermission('MANAGE_ROLES_OR_PERMISSIONS')) {
+        msg.delete();
+        msg.reply('I\'m trying to eat eucalyptus in peace, so please stop mentioning everyone.');
+      }
     }
 };
